refactor export request handler
